+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * **Fix Kafka source connection problems**
 * **Fix Azure AI Search session handling**
+* **Properly log elasticsearch upload errors**
 
 ### Enhancements